+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Add support for initializing fixed-size arrays inside kernels using `wp.zeros()`
   ([GH-794](https://github.com/NVIDIA/warp/issues/794)).
 - Add support for `IntEnum` and `IntFlag` inside Warp kernels ([GH-529](https://github.com/NVIDIA/warp/issues/529)).
+- Add `wp.atomic_and()` (`&=`), `wp.atomic_or()` (`|=`), and `wp.atomic_xor()` (`^=`) built-ins for atomic bitwise operations
+  ([GH-886](https://github.com/NVIDIA/warp/issues/886)).
 
 ### Changed